+++ apps/playground/src/components/test1.vue
@@ -12,6 +12,9 @@
   <div v-for="i in 2" :key="i" style="width:2000px;">
     test-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxxtest-1xxxx
   </div>
+  <ModalFooter>
+    123123xxx
+  </ModalFooter>
 </template>
 
 <style scoped>
@@ -21,7 +24,7 @@
 <script setup lang="ts">
 import { reactive } from 'vue'
 import { Form as AForm, type FieldRule } from '@arco-design/web-vue'
-import { onSubmit, useModal } from '@gopowerteam/modal-render'
+import { ModalFooter, onSubmit, useModal } from '@gopowerteam/modal-render'
 
 const modal = useModal()
 const formModel = reactive({