+++ packages/modal-render/src/index.ts
@@ -1,6 +1,7 @@
 import 'virtual:uno.css'
 import ModalProvider from './components/modal-provider.vue'
 import ModalContainer from './components/modal-container.vue'
+import ModalFooter from './components/modal-footer.vue'
 import { useModal } from './hooks/use-modal'
 import { ModalKey } from './constants'
 
@@ -10,6 +11,7 @@ export * from './interfaces'
 export {
   ModalContainer,
   ModalProvider,
+  ModalFooter,
   useModal,
   ModalKey,
 }