+++ packages/modal-render/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gopowerteam/modal-render
 
+## 0.0.24
+
+### Patch Changes
+
+- add modal footer
+
 ## 0.0.23
 
 ### Patch Changes