+++ packages/table-render/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gopowerteam/table-render
 
+## 0.0.74
+
+### Patch Changes
+
+- Updated dependencies
+  - @gopowerteam/modal-render@0.0.24
+
 ## 0.0.73
 
 ### Patch Changes